Move fixture into test
